fix(SRC): 👾  use size_t -> unsigned int

+++ ex01/include/Span.hpp
@@ -26,7 +26,7 @@ class Span {
 
   template <typename Iterator>
   void addNumbers(Iterator begin, Iterator end) {
-    size_t distance = std::distance(begin, end);
+    unsigned int distance = std::distance(begin, end);
     if (_numbers.size() + distance > _maxSize) {
       throw std::runtime_error("Cannot add range: would exceed maximum size");
     }
